refactor(record): allow to init record internally
